# Use Bumpalo for rust single threaded