--- a/rust/Cargo.toml
+++ b/rust/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bumpalo = { version = "3.14.0", features = ["collections"] }
 mimalloc = "0.1.39"
 rustc_data_structures = "0.0.1"
 serde = { version = "1.0.188", features = ["derive"] }