Before

Benchmark 1: ./target/release/rust
Processing time (w/o IO): 23.382257ms
Processing time (w/o IO): 23.112573ms
Processing time (w/o IO): 23.295275ms
Processing time (w/o IO): 23.196617ms
Processing time (w/o IO): 23.227118ms
Processing time (w/o IO): 22.486412ms
Processing time (w/o IO): 22.52874ms
Processing time (w/o IO): 22.337788ms
Processing time (w/o IO): 22.548885ms
Processing time (w/o IO): 22.609888ms
Processing time (w/o IO): 22.483896ms
Processing time (w/o IO): 22.613415ms
Processing time (w/o IO): 23.129902ms
Processing time (w/o IO): 22.993105ms
Processing time (w/o IO): 22.927814ms
  Time (mean ± σ):      31.4 ms ±   0.7 ms    [User: 26.4 ms, System: 4.8 ms]
  Range (min … max):    30.7 ms …  32.9 ms    10 runs

After

Benchmark 1: ./target/release/rust
Processing time (w/o IO): 12.154076ms
Processing time (w/o IO): 13.217269ms
Processing time (w/o IO): 12.083854ms
Processing time (w/o IO): 12.105984ms
Processing time (w/o IO): 12.066213ms
Processing time (w/o IO): 12.018752ms
Processing time (w/o IO): 12.074031ms
Processing time (w/o IO): 12.183986ms
Processing time (w/o IO): 12.199131ms
Processing time (w/o IO): 12.294161ms
Processing time (w/o IO): 12.168269ms
Processing time (w/o IO): 12.27678ms
Processing time (w/o IO): 12.455885ms
Processing time (w/o IO): 12.238231ms
Processing time (w/o IO): 12.197747ms
  Time (mean ± σ):      21.0 ms ±   0.3 ms    [User: 16.8 ms, System: 4.1 ms]
  Range (min … max):    20.7 ms …  21.9 ms    10 runs