@masmullin2000 2023-10-06
This pull request has 2 independent commits.  First is for single threaded rust.  Second is for rust concurrent.  Second commit requires rust 1.73 (released Oct 5, 2023)