@jinyus 2023-10-06
The github action should grab the latest stable version right? I don't want the build to fail in the workflow.